requirements.txt
pytesseract
opencv-python
pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
tesseract-ocr-por